Support Android OTA signing schemes.
This adds support for OTA signing for Android A/B devices. Specifically adds
code that knows how to parse and sign Android verified-boot images, and also
knows how to sign dm-verity (non-boot) partitions and Brillo A/B-system
payloads. The latter two merely sign with RSA: the logic for constructing the
bits to be signed remains on the build server side.